pyproject.toml
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "aioresponses",
    "pytest-homeassistant-custom-component",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["custom_components"]

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = [".venv", ".claude", "tests/__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]  # line length already enforced via line-length

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["B", "UP", "F401", "F841", "E402"]  # tests can use older idioms / bare assert patterns

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true
follow_imports = "silent"
exclude = [
    "^tests/",
    "^\\.venv/",
    "^\\.claude/",
]

[[tool.mypy.overrides]]
module = "shop2parcel.options_flow"
ignore_errors = true

[[tool.mypy.overrides]]
module = "shop2parcel.config_flow"
ignore_errors = true

[[tool.mypy.overrides]]
module = "shop2parcel.coordinator"
ignore_errors = true
